requirements.txt
cohere==4.20.0
gTTS==2.3.2
langchain==0.0.264
pymongo==4.4.1
python-dotenv==1.0.0
streamlit==1.24.1
